+++ CHANGELOG.md
@@ -9,8 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+-   Added new dependency `LibMobs`.
 -   Added options to time datatext to display time as 24-hour or 12-hour format.
--   New dependency LibMobs.
+-   Added dungeon portal shortcut inside challenge frame.
 
 ### Changed